Small widgets created for navigation screen
