+++ lib/src/map/navigation/widgets/simulate_btn_widget.dart
@@ -1,0 +1,32 @@
+import 'package:flutter/material.dart';
+import 'package:graphhooper_route_navigation/src/map/navigation/providers/map_controller_provider.dart';
+
+class SimulateButton extends StatefulWidget {
+  const SimulateButton({
+    super.key,
+  });
+
+  @override
+  State<SimulateButton> createState() => _SimulateButtonState();
+}
+
+class _SimulateButtonState extends State<SimulateButton> {
+  bool isSimulateRouting = false;
+
+  @override
+  Widget build(BuildContext context) {
+    final mapController = MapControllerProvider.of(context);
+    return SizedBox(
+      height: 30.0,
+      child: MaterialButton(
+          child: const Text(
+            'Simulate',
+            style: TextStyle(color: Colors.black),
+          ),
+          onPressed: () {
+            isSimulateRouting = !isSimulateRouting;
+            mapController.simulateRouting();
+          }),
+    );
+  }
+}
